Minor fix after rebasing to master and move job tests into own package
