+++ xyz-hub-test/src/test/java/com/here/xyz/hub/rest/jobs/JobApiImportIT.java
@@ -16,7 +16,7 @@
  * SPDX-License-Identifier: Apache-2.0
  * License-Filename: LICENSE
  */
-package com.here.xyz.hub.rest;
+package com.here.xyz.hub.rest.jobs;
 
 import static com.here.xyz.hub.rest.Api.HeaderValues.APPLICATION_GEO_JSON;
 import static com.here.xyz.hub.rest.Api.HeaderValues.APPLICATION_JSON;
@@ -30,6 +30,9 @@ import static org.junit.Assert.assertTrue;
 import com.here.xyz.httpconnector.util.jobs.Import;
 import com.here.xyz.httpconnector.util.jobs.ImportObject;
 import com.here.xyz.httpconnector.util.jobs.Job;
+import com.here.xyz.hub.auth.TestAuthenticator;
+import com.here.xyz.hub.rest.TestSpaceWithFeature;
+import com.here.xyz.hub.rest.TestWithSpaceCleanup;
 import io.vertx.core.json.JsonArray;
 import io.vertx.core.json.JsonObject;
 import java.net.URL;
@@ -56,16 +59,16 @@ public class JobApiImportIT extends JobApiIT {
     @Before
     public void setupAndCleanBefore(){
         deleteAllJobsOnSpace(getScopedSpaceId(testSpaceId1, scope));
-        removeSpace(getScopedSpaceId(testSpaceId1, scope));
+        TestWithSpaceCleanup.removeSpace(getScopedSpaceId(testSpaceId1, scope));
 
         /** Create empty test space */
-        createSpaceWithCustomStorage(getScopedSpaceId(testSpaceId1, scope), "psql", null);
+        TestSpaceWithFeature.createSpaceWithCustomStorage(getScopedSpaceId(testSpaceId1, scope), "psql", null);
     }
 
     @After
     public void cleanAfter(){
         deleteAllJobsOnSpace(getScopedSpaceId(testSpaceId1, scope));
-        removeSpace(getScopedSpaceId(testSpaceId1, scope));
+        TestWithSpaceCleanup.removeSpace(getScopedSpaceId(testSpaceId1, scope));
     }
 
     @Test
@@ -77,7 +80,7 @@ public class JobApiImportIT extends JobApiIT {
         String resp = given()
                 .accept(APPLICATION_JSON)
                 .contentType(APPLICATION_JSON)
-                .headers(getAuthHeaders(AuthProfile.ACCESS_ALL))
+                .headers(TestAuthenticator.getAuthHeaders(AuthProfile.ACCESS_ALL))
                 .post("/spaces/" + getScopedSpaceId(testSpaceId1, scope) + "/job/"+job.getId()+"/execute?command=createUploadUrl")
                 .getBody().asString();
         String url = new JsonObject(resp).getString("url");
@@ -89,7 +92,7 @@ public class JobApiImportIT extends JobApiIT {
         given()
                 .accept(APPLICATION_JSON)
                 .contentType(APPLICATION_JSON)
-                .headers(getAuthHeaders(AuthProfile.ACCESS_ALL))
+                .headers(TestAuthenticator.getAuthHeaders(AuthProfile.ACCESS_ALL))
                 .post("/spaces/" + getScopedSpaceId(testSpaceId1, scope) + "/job/"+job.getId()+"/execute?command=start")
                 .then()
                 .statusCode(NO_CONTENT.code());
@@ -101,7 +104,7 @@ public class JobApiImportIT extends JobApiIT {
         given()
                 .accept(APPLICATION_GEO_JSON)
                 .contentType(APPLICATION_GEO_JSON)
-                .headers(getAuthHeaders(AuthProfile.ACCESS_OWNER_1_ADMIN))
+                .headers(TestAuthenticator.getAuthHeaders(AuthProfile.ACCESS_OWNER_1_ADMIN))
                 .when()
                 .get("/spaces/" + getScopedSpaceId(testSpaceId1, scope) +"/search")
                 .then()
@@ -119,7 +122,7 @@ public class JobApiImportIT extends JobApiIT {
             String resp = given()
                     .accept(APPLICATION_JSON)
                     .contentType(APPLICATION_JSON)
-                    .headers(getAuthHeaders(AuthProfile.ACCESS_ALL))
+                    .headers(TestAuthenticator.getAuthHeaders(AuthProfile.ACCESS_ALL))
                     .post("/spaces/" + getScopedSpaceId(testSpaceId1, scope) + "/job/" + job.getId() + "/execute?command=createUploadUrl")
                     .getBody().asString();
             String url = new JsonObject(resp).getString("url");
@@ -132,7 +135,7 @@ public class JobApiImportIT extends JobApiIT {
         given()
                 .accept(APPLICATION_JSON)
                 .contentType(APPLICATION_JSON)
-                .headers(getAuthHeaders(AuthProfile.ACCESS_ALL))
+                .headers(TestAuthenticator.getAuthHeaders(AuthProfile.ACCESS_ALL))
                 .post("/spaces/" + getScopedSpaceId(testSpaceId1, scope) + "/job/"+job.getId()+"/execute?command=start")
                 .then()
                 .statusCode(NO_CONTENT.code());
@@ -144,7 +147,7 @@ public class JobApiImportIT extends JobApiIT {
         given()
                 .accept(APPLICATION_GEO_JSON)
                 .contentType(APPLICATION_GEO_JSON)
-                .headers(getAuthHeaders(AuthProfile.ACCESS_OWNER_1_ADMIN))
+                .headers(TestAuthenticator.getAuthHeaders(AuthProfile.ACCESS_OWNER_1_ADMIN))
                 .when()
                 .get("/spaces/" + getScopedSpaceId(testSpaceId1, scope) +"/search")
                 .then()
@@ -181,7 +184,7 @@ public class JobApiImportIT extends JobApiIT {
         String resp = given()
                 .accept(APPLICATION_JSON)
                 .contentType(APPLICATION_JSON)
-                .headers(getAuthHeaders(AuthProfile.ACCESS_ALL))
+                .headers(TestAuthenticator.getAuthHeaders(AuthProfile.ACCESS_ALL))
                 .post("/spaces/" + getScopedSpaceId(testSpaceId1, scope) + "/job/"+job.getId()+"/execute?command=createUploadUrl")
                 .getBody().asString();
         String url = new JsonObject(resp).getString("url");
@@ -193,7 +196,7 @@ public class JobApiImportIT extends JobApiIT {
         given()
                 .accept(APPLICATION_JSON)
                 .contentType(APPLICATION_JSON)
-                .headers(getAuthHeaders(AuthProfile.ACCESS_ALL))
+                .headers(TestAuthenticator.getAuthHeaders(AuthProfile.ACCESS_ALL))
                 .post("/spaces/" + getScopedSpaceId(testSpaceId1, scope) + "/job/"+job.getId()+"/execute?command=start")
                 .then()
                 .statusCode(NO_CONTENT.code());
@@ -205,7 +208,7 @@ public class JobApiImportIT extends JobApiIT {
         given()
                 .accept(APPLICATION_GEO_JSON)
                 .contentType(APPLICATION_GEO_JSON)
-                .headers(getAuthHeaders(AuthProfile.ACCESS_OWNER_1_ADMIN))
+                .headers(TestAuthenticator.getAuthHeaders(AuthProfile.ACCESS_OWNER_1_ADMIN))
                 .when()
                 .get("/spaces/" + getScopedSpaceId(testSpaceId1, scope) +"/search")
                 .then()
@@ -222,7 +225,7 @@ public class JobApiImportIT extends JobApiIT {
         String resp1 = given()
                 .accept(APPLICATION_JSON)
                 .contentType(APPLICATION_JSON)
-                .headers(getAuthHeaders(AuthProfile.ACCESS_ALL))
+                .headers(TestAuthenticator.getAuthHeaders(AuthProfile.ACCESS_ALL))
                 .post("/spaces/" + getScopedSpaceId(testSpaceId1, scope) + "/job/"+job.getId()+"/execute?command=createUploadUrl")
                 .getBody().asString();
         String url = new JsonObject(resp1).getString("url");
@@ -233,7 +236,7 @@ public class JobApiImportIT extends JobApiIT {
         String resp2 = given()
                 .accept(APPLICATION_JSON)
                 .contentType(APPLICATION_JSON)
-                .headers(getAuthHeaders(AuthProfile.ACCESS_ALL))
+                .headers(TestAuthenticator.getAuthHeaders(AuthProfile.ACCESS_ALL))
                 .post("/spaces/" + getScopedSpaceId(testSpaceId1, scope) + "/job/"+job.getId()+"/execute?command=createUploadUrl&urlCount="+urlCount)
                 .getBody().asString();
         JsonArray urls = new JsonObject(resp2).getJsonArray("urls");
@@ -246,7 +249,7 @@ public class JobApiImportIT extends JobApiIT {
         String resp3 = given()
                 .accept(APPLICATION_JSON)
                 .contentType(APPLICATION_JSON)
-                .headers(getAuthHeaders(AuthProfile.ACCESS_ALL))
+                .headers(TestAuthenticator.getAuthHeaders(AuthProfile.ACCESS_ALL))
                 .post("/spaces/" + getScopedSpaceId(testSpaceId1, scope) + "/job/"+job.getId()+"/execute?command=createUploadUrl")
                 .getBody().asString();
         urls = new JsonObject(resp3).getJsonArray("urls");
@@ -257,7 +260,7 @@ public class JobApiImportIT extends JobApiIT {
         given()
                 .accept(APPLICATION_JSON)
                 .contentType(APPLICATION_JSON)
-                .headers(getAuthHeaders(AuthProfile.ACCESS_ALL))
+                .headers(TestAuthenticator.getAuthHeaders(AuthProfile.ACCESS_ALL))
                 .get("/spaces/" + getScopedSpaceId(testSpaceId1, scope) + "/job/"+job.getId())
                 .then()
                 .statusCode(OK.code())
@@ -268,7 +271,7 @@ public class JobApiImportIT extends JobApiIT {
         given()
                 .accept(APPLICATION_JSON)
                 .contentType(APPLICATION_JSON)
-                .headers(getAuthHeaders(AuthProfile.ACCESS_ALL))
+                .headers(TestAuthenticator.getAuthHeaders(AuthProfile.ACCESS_ALL))
                 .post("/spaces/" + getScopedSpaceId(testSpaceId1, scope) + "/job/"+job.getId()+"/execute?command=start")
                 .then()
                 .statusCode(NO_CONTENT.code());
@@ -280,7 +283,7 @@ public class JobApiImportIT extends JobApiIT {
         given()
                 .accept(APPLICATION_GEO_JSON)
                 .contentType(APPLICATION_GEO_JSON)
-                .headers(getAuthHeaders(AuthProfile.ACCESS_OWNER_1_ADMIN))
+                .headers(TestAuthenticator.getAuthHeaders(AuthProfile.ACCESS_OWNER_1_ADMIN))
                 .when()
                 .get("/spaces/" + getScopedSpaceId(testSpaceId1, scope) +"/search")
                 .then()
@@ -296,7 +299,7 @@ public class JobApiImportIT extends JobApiIT {
         String resp = given()
                 .accept(APPLICATION_JSON)
                 .contentType(APPLICATION_JSON)
-                .headers(getAuthHeaders(AuthProfile.ACCESS_ALL))
+                .headers(TestAuthenticator.getAuthHeaders(AuthProfile.ACCESS_ALL))
                 .post("/spaces/" + getScopedSpaceId(testSpaceId1, scope) + "/job/"+job.getId()+"/execute?command=createUploadUrl")
                 .getBody().asString();
         String url = new JsonObject(resp).getString("url");
@@ -308,7 +311,7 @@ public class JobApiImportIT extends JobApiIT {
         given()
                 .accept(APPLICATION_JSON)
                 .contentType(APPLICATION_JSON)
-                .headers(getAuthHeaders(AuthProfile.ACCESS_ALL))
+                .headers(TestAuthenticator.getAuthHeaders(AuthProfile.ACCESS_ALL))
                 .post("/spaces/" + getScopedSpaceId(testSpaceId1, scope) + "/job/"+job.getId()+"/execute?command=start")
                 .then()
                 .statusCode(NO_CONTENT.code());
@@ -323,7 +326,7 @@ public class JobApiImportIT extends JobApiIT {
         given()
                 .accept(APPLICATION_JSON)
                 .contentType(APPLICATION_JSON)
-                .headers(getAuthHeaders(AuthProfile.ACCESS_ALL))
+                .headers(TestAuthenticator.getAuthHeaders(AuthProfile.ACCESS_ALL))
                 .post("/spaces/" + getScopedSpaceId(testSpaceId1, scope) + "/job/"+job.getId()+"/execute?command=retry")
                 .then()
                 .statusCode(NO_CONTENT.code());
@@ -338,7 +341,7 @@ public class JobApiImportIT extends JobApiIT {
         given()
                 .accept(APPLICATION_JSON)
                 .contentType(APPLICATION_JSON)
-                .headers(getAuthHeaders(AuthProfile.ACCESS_ALL))
+                .headers(TestAuthenticator.getAuthHeaders(AuthProfile.ACCESS_ALL))
                 .post("/spaces/" + getScopedSpaceId(testSpaceId1, scope) + "/job/"+job.getId()+"/execute?command=retry")
                 .then()
                 .statusCode(NO_CONTENT.code());
@@ -355,7 +358,7 @@ public class JobApiImportIT extends JobApiIT {
         given()
                 .accept(APPLICATION_JSON)
                 .contentType(APPLICATION_JSON)
-                .headers(getAuthHeaders(AuthProfile.ACCESS_ALL))
+                .headers(TestAuthenticator.getAuthHeaders(AuthProfile.ACCESS_ALL))
                 .post("/spaces/" + getScopedSpaceId(testSpaceId1, scope) + "/job/"+job.getId()+"/execute?command=start")
                 .then()
                 .statusCode(NO_CONTENT.code());
@@ -380,7 +383,7 @@ public class JobApiImportIT extends JobApiIT {
             String resp = given()
                     .accept(APPLICATION_JSON)
                     .contentType(APPLICATION_JSON)
-                    .headers(getAuthHeaders(AuthProfile.ACCESS_ALL))
+                    .headers(TestAuthenticator.getAuthHeaders(AuthProfile.ACCESS_ALL))
                     .post("/spaces/" + getScopedSpaceId(testSpaceId1, scope) + "/job/"+job.getId()+"/execute?command=createUploadUrl")
                     .getBody().asString();
             String url = new JsonObject(resp).getString("url");
@@ -393,7 +396,7 @@ public class JobApiImportIT extends JobApiIT {
         given()
                 .accept(APPLICATION_JSON)
                 .contentType(APPLICATION_JSON)
-                .headers(getAuthHeaders(AuthProfile.ACCESS_ALL))
+                .headers(TestAuthenticator.getAuthHeaders(AuthProfile.ACCESS_ALL))
                 .post("/spaces/" + getScopedSpaceId(testSpaceId1, scope) + "/job/"+job.getId()+"/execute?command=start")
                 .then()
                 .statusCode(NO_CONTENT.code());
@@ -405,7 +408,7 @@ public class JobApiImportIT extends JobApiIT {
         given()
                 .accept(APPLICATION_GEO_JSON)
                 .contentType(APPLICATION_GEO_JSON)
-                .headers(getAuthHeaders(AuthProfile.ACCESS_OWNER_1_ADMIN))
+                .headers(TestAuthenticator.getAuthHeaders(AuthProfile.ACCESS_OWNER_1_ADMIN))
                 .when()
                 .get("/spaces/" + getScopedSpaceId(testSpaceId1, scope) +"/search")
                 .then()
@@ -417,16 +420,16 @@ public class JobApiImportIT extends JobApiIT {
         String newSpace = "new_empty_space";
         String importId = "manual_big_import";
 
-        removeSpace(newSpace);
+        TestWithSpaceCleanup.removeSpace(newSpace);
         deleteAllJobsOnSpace(newSpace);
-        createSpaceWithCustomStorage(newSpace, "psql", null);
+        TestSpaceWithFeature.createSpaceWithCustomStorage(newSpace, "psql", null);
         /** Create job */
         Job job = createTestJobWithId(newSpace, importId, JobApiIT.Type.Import, Job.CSVFormat.JSON_WKB);
 
         String resp = given()
                 .accept(APPLICATION_JSON)
                 .contentType(APPLICATION_JSON)
-                .headers(getAuthHeaders(AuthProfile.ACCESS_ALL))
+                .headers(TestAuthenticator.getAuthHeaders(AuthProfile.ACCESS_ALL))
                 .post("/spaces/" + newSpace + "/job/"+job.getId()+"/execute?command=createUploadUrl")
                 .getBody().asString();
         String url = new JsonObject(resp).getString("url");
@@ -438,7 +441,7 @@ public class JobApiImportIT extends JobApiIT {
         resp = given()
                 .accept(APPLICATION_JSON)
                 .contentType(APPLICATION_JSON)
-                .headers(getAuthHeaders(AuthProfile.ACCESS_ALL))
+                .headers(TestAuthenticator.getAuthHeaders(AuthProfile.ACCESS_ALL))
                 .post("/spaces/" + newSpace + "/job/"+job.getId()+"/execute?command=createUploadUrl")
                 .getBody().asString();
         url = new JsonObject(resp).getString("url");
@@ -450,7 +453,7 @@ public class JobApiImportIT extends JobApiIT {
         given()
                 .accept(APPLICATION_JSON)
                 .contentType(APPLICATION_JSON)
-                .headers(getAuthHeaders(AuthProfile.ACCESS_ALL))
+                .headers(TestAuthenticator.getAuthHeaders(AuthProfile.ACCESS_ALL))
                 .post("/spaces/" + newSpace + "/job/"+job.getId()+"/execute?command=start")
                 .then()
                 .statusCode(NO_CONTENT.code());
@@ -459,6 +462,6 @@ public class JobApiImportIT extends JobApiIT {
         pollStatus(newSpace, job.getId(), Job.Status.finalized, Job.Status.failed);
 
         deleteAllJobsOnSpace(newSpace);
-        removeSpace(newSpace);
+        TestWithSpaceCleanup.removeSpace(newSpace);
     }
 }
